requirements.txt
pybit==5.8.0
aiohttp>=3.9.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
